# docs/swagger.yaml
basePath: /
definitions:
...
  crud.ValidationErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  crud.ValidationErrorResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/crud.ValidationErrorDetail'
        type: array
      error:
        type: string
    type: object
...
  models.CheckoutForm:
    properties:
      address:
        type: string
      bin:
        type: string
      comment:
        type: string
      companyName:
        type: string
      customerType:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  models.Manufacturer:
    properties:
      id:
        type: string
      logo:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Noble Group Services e-commerce platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Noble Group Services API
  version: "1.0"
paths:
...
  /orders:
    post:
      consumes:
      - application/json
      description: Place a new order with the items in the cart
      parameters:
      - description: Checkout Form
        in: body
        name: checkoutForm
        required: true
        schema:
          $ref: '#/definitions/models.CheckoutForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/crud.ValidationErrorResponse'
      summary: Place a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      description: Delete an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Order not found
          schema:
            type: string
      summary: Delete order
      tags:
      - orders
...
  /products/manufacturers:
    get:
      description: Get a list of all manufacturers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Manufacturer'
            type: array
      summary: Get all manufacturers
      tags:
      - manufacturers
    post:
      consumes:
      - application/json
      description: Create a new manufacturer
      parameters:
      - description: Manufacturer
        in: body
        name: manufacturer
        required: true
        schema:
          $ref: '#/definitions/models.Manufacturer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Manufacturer'
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create a manufacturer
      tags:
      - manufacturers
  /products/manufacturers/{id}:
    delete:
      description: Delete a manufacturer
      parameters:
      - description: Manufacturer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Manufacturer not found
          schema:
            type: string
      summary: Delete manufacturer
      tags:
      - manufacturers
    get:
      description: Get details of a specific manufacturer
      parameters:
      - description: Manufacturer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Manufacturer'
        "404":
          description: Manufacturer not found
          schema:
            type: string
      summary: Get manufacturer by ID
      tags:
      - manufacturers
    put:
      consumes:
      - application/json
      description: Update an existing manufacturer
      parameters:
      - description: Manufacturer ID
        in: path
        name: id
        required: true
        type: string
      - description: Manufacturer
        in: body
        name: manufacturer
        required: true
        schema:
          $ref: '#/definitions/models.Manufacturer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Manufacturer'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Manufacturer not found
          schema:
            type: string
      summary: Update manufacturer
      tags:
      - manufacturers
swagger: "2.0"